Build out categoies page
A lot happened this commit. Building the SQL queries, rendering from the controller with the query data, creating the view with the category partial.

A big obstacle was figuring out how to make the page not look so bland. I tinkered around with the main section background color, alternating card background colors, borders, etc. But nothing was feeling right, and I didn't want to hard code anything considering the categories would be dynamic.

Eventually, I just decided to add a column to the database for each category's color value. When adding/updating a category, users can now assign a color.

This really helped bring life to the page. Next steps are building the products page, along with the filtering feature.