Add basic sorting functionality
Sorting consists of no sorting, high to low price, and low to high price.

The db query was updated by dynamically adding the sort statement depending on the form select input query parameter.

In the products view, the correct dropdown option was selected by using the ternary operator to add 'selected' to the select option matching the sort value.