Finish add category functionality
Upon form submission, the req.body parameters are validated and sanitized before being sent to the
addCategory query. To ensure that duplicate category names are avoided, category_name was made UNIQUE,
and the query call was placed in a try/catch block. If an error is caught, then the form is re-rendered
with an error saying that the category already exists.

An important note is that I used a custom sanitizer to double any occurences of apostrophe's so they don't
break our SQL queries. I had though about escaping, but since it is not the best practice, I opted for this
other solution.