Complete addProduct functionality
This involved full validation/santization of form input.

Brand name had to be queried with the brand table to see if the brand existed. If so,
the brand_id was retrieved and used when inserting the product into the product table.
If not, the brand was inserted into the brand table and the new brand_id was used.

SQL errors were caught with try/catch blocks.