Set up basic update routing and populate form with product data
Using route parameters, we fetch the product data and use that to populate
the update form values. Styling is consistent with the add product form.

This is routed to the "/update/:product_id" POST route, which will validate/sanitize
inputs and proceed to query the database to update the product_id information.

Next steps: complete updateProductPost controller method, write updateProduct db query.